[PATCH] linux: support unpacked source trees by Mike Frysinger
I keep the kernel sources for my target laying around on my system
(since that's what I'm developing against), so having to manage a
tarball with it is a pita.  So add a new "local tree" option so that
people can point the build system to that.

While I could have this new option "unpack" the tree into the br
tree by doing a recursive cp or rsync, it seems like it makes more
sense to support out-of-tree building (O=).  So split the linux.mk
logic to have a "source" and "build" dir.

http://permalink.gmane.org/gmane.comp.lib.uclibc.buildroot/26358